+++ src/main/java/pairmatching/Application.java
@@ -5,6 +5,8 @@ import pairmatching.domain.PairKey;
 import pairmatching.repository.PairRepository;
 import pairmatching.util.Command;
 import pairmatching.util.FeatureMenu;
+import pairmatching.util.Initiator;
+import pairmatching.util.PairMatcher;
 import pairmatching.view.InputView;
 import pairmatching.view.OutputView;
