+++ src/main/java/pairmatching/util/PairMatcher.java
@@ -1,4 +1,4 @@
-package pairmatching;
+package pairmatching.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
